Removed CollectionUtils
