+++ selcukes-databind/src/main/java/io/github/selcukes/databind/utils/Streams.java
@@ -35,8 +35,8 @@ public class Streams {
     /**
      * It takes an iterator and returns a stream.
      *
-     * @param iterator The iterator to convert to a stream.
-     * @return A stream of the iterator.
+     * @param  iterator The iterator to convert to a stream.
+     * @return          A stream of the iterator.
      */
     public <T> Stream<T> of(final Iterator<? extends T> iterator) {
         return StreamSupport
@@ -44,14 +44,16 @@ public class Streams {
     }
 
     /**
-     * "Return the index of the first element in the list that matches the predicate, or an empty OptionalInt if no such
-     * element exists."
+     * "Return the index of the first element in the list that matches the
+     * predicate, or an empty OptionalInt if no such element exists."
+     * <p>
+     * The first thing we do is convert the list to a Stream. This is done using
+     * the of() method
      *
-     * The first thing we do is convert the list to a Stream. This is done using the of() method
-     *
-     * @param elements The list of elements to search through.
-     * @param predicate A function that takes an element of the list and returns a boolean.
-     * @return OptionalInt
+     * @param  elements  The list of elements to search through.
+     * @param  predicate A function that takes an element of the list and
+     *                   returns a boolean.
+     * @return           OptionalInt
      */
     public <T> OptionalInt indexOf(List<T> elements, IntPredicate predicate) {
         return of(elements)
@@ -62,8 +64,8 @@ public class Streams {
     /**
      * It returns an IntStream of the indices of the elements in the given list.
      *
-     * @param elements The list of elements to iterate over.
-     * @return An IntStream of the indexes of the elements in the list.
+     * @param  elements The list of elements to iterate over.
+     * @return          An IntStream of the indexes of the elements in the list.
      */
     public <T> IntStream of(List<T> elements) {
         return of(0, elements.size());
@@ -72,9 +74,9 @@ public class Streams {
     /**
      * Returns an IntStream of the numbers between start and end, inclusive.
      *
-     * @param start The starting value of the range.
-     * @param end   The end value (exclusive) for the range to be created
-     * @return IntStream
+     * @param  start The starting value of the range.
+     * @param  end   The end value (exclusive) for the range to be created
+     * @return       IntStream
      */
     public <T> IntStream of(int start, int end) {
         return IntStream.range(start, end);
@@ -84,8 +86,8 @@ public class Streams {
      * Skip the first row, then for each row, create a map from the headers to
      * the values.
      *
-     * @param cells The list of lists of strings that represent the table.
-     * @return A list of maps.
+     * @param  cells The list of lists of strings that represent the table.
+     * @return       A list of maps.
      */
     public List<Map<String, String>> listOfMap(List<List<String>> cells) {
         var headers = cells.get(0);
@@ -96,18 +98,50 @@ public class Streams {
     /**
      * It takes a list of maps and returns a map of lists
      *
-     * @param listMap The list of maps to be converted.
-     * @return A map of lists.
+     * @param  listMap The list of maps to be converted.
+     * @return         A map of lists.
      */
     public <K, V> Map<K, List<V>> mapOfList(List<Map<K, V>> listMap) {
         return listMap.stream()
                 .flatMap(map -> map.entrySet().stream())
                 .collect(Collectors.groupingBy(
-                        Map.Entry::getKey,
-                        Collectors.mapping(
-                                Map.Entry::getValue,
-                                Collectors.toList()
-                        )
-                ));
+                    Map.Entry::getKey,
+                    Collectors.mapping(
+                        Map.Entry::getValue,
+                        Collectors.toList())));
+    }
+
+    /**
+     * Convert a list of lists of strings into a 2D array of strings.
+     *
+     * @param  cells The list of lists of strings to convert to a 2D array.
+     * @return       A 2D array of Strings
+     */
+    public String[][] toArray(final List<List<String>> cells) {
+        return cells.stream()
+                .map(row -> row.toArray(String[]::new))
+                .toArray(String[][]::new);
+    }
+
+    /**
+     * Return a list of trimmed strings from the given list of strings.
+     *
+     * @param  list The list to trim.
+     * @return      A list of strings that have been trimmed.
+     */
+    public List<String> trim(final List<String> list) {
+        return list.stream()
+                .map(String::trim)
+                .collect(Collectors.toList());
+    }
+
+    /**
+     * It takes an enum class and returns a stream of the enum's string values.
+     *
+     * @param  enumData The enum class to get the values from.
+     * @return          A stream of strings.
+     */
+    public Stream<String> of(final Class<? extends Enum<?>> enumData) {
+        return Stream.of(enumData.getEnumConstants()).map(Enum::toString);
     }
 }
