+++ selcukes-databind/src/test/java/io/github/selcukes/databind/tests/CsvTest.java
@@ -28,7 +28,7 @@ public class CsvTest {
     @Test
     public void csvDataReaderTest() {
         Path filePath = Path.of(
-                Objects.requireNonNull(CsvTest.class.getClassLoader().getResource("employee.csv")).toURI());
+            Objects.requireNonNull(CsvTest.class.getClassLoader().getResource("employee.csv")).toURI());
         var csvData = CsvMapper.parse(filePath);
 
         csvData.forEach(map -> map.computeIfPresent("ID", (k, v) -> {
